requirements.txt
streamlit              # Interactive data app framework  
langchain              # LangChain framework for building language model applications  
langchain-core         # Core utilities and abstractions for LangChain  
langchain-community    # Community-contributed integrations and tools for LangChain  

langchain-groq         # Groq integration for LangChain  
python-dotenv          # Load environment variables from .env files  

langchain-huggingface
langchain-chroma
pysqlite3-binary
